Initial commit of plist based releases coverflow and album info

+++ Classes/ReleasesView.h
@@ -17,7 +17,7 @@
 	UIButton *infoButton;
 	
 	NSMutableArray *covers; // album covers images
-	NSInteger coverIndex;
+	int coverIndex;
 }
 
 @property (retain,nonatomic) TKCoverflowView *coverflow;
